Finalizing Performance optimization.
