+++ src/ByteAether.Ulid.sln
@@ -18,8 +18,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ByteAether.Ulid.Tests.AotConsole", "ByteAether.Ulid.Tests.AotConsole\ByteAether.Ulid.Tests.AotConsole.csproj", "{A16BAD2D-8F7F-49E5-880A-349E31F6E70E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ByteAether.Ulid2", "ByteAether.Ulid2\ByteAether.Ulid2.csproj", "{8B5C57B8-11A3-4D0B-A444-9BAC50A339B6}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -42,10 +40,6 @@ Global
 		{A16BAD2D-8F7F-49E5-880A-349E31F6E70E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A16BAD2D-8F7F-49E5-880A-349E31F6E70E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A16BAD2D-8F7F-49E5-880A-349E31F6E70E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8B5C57B8-11A3-4D0B-A444-9BAC50A339B6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8B5C57B8-11A3-4D0B-A444-9BAC50A339B6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8B5C57B8-11A3-4D0B-A444-9BAC50A339B6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8B5C57B8-11A3-4D0B-A444-9BAC50A339B6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
