--- src/ByteAether.Ulid2/Compatibility/Range.cs
@@ -1,141 +1,0 @@
-﻿#if NETSTANDARD2_0
-
-// ReSharper disable All
-#pragma warning disable
-// https://github.com/dotnet/runtime/blob/v9.0.0/src/libraries/System.Private.CoreLib/src/System/Range.cs
-
-using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
-using System.Runtime.CompilerServices;
-
-namespace System
-{
-	/// <summary>Represent a range has start and end indexes.</summary>
-	/// <remarks>
-	/// Range is used by the C# compiler to support the range syntax.
-	/// <code>
-	/// int[] someArray = new int[5] { 1, 2, 3, 4, 5 };
-	/// int[] subArray1 = someArray[0..2]; // { 1, 2 }
-	/// int[] subArray2 = someArray[1..^0]; // { 2, 3, 4, 5 }
-	/// </code>
-	/// </remarks>
-#if SYSTEM_PRIVATE_CORELIB || MICROSOFT_BCL_MEMORY
-    public
-#else
-	internal
-#endif
-	readonly struct Range : IEquatable<Range>
-	{
-		/// <summary>Represent the inclusive start index of the Range.</summary>
-		public Index Start { get; }
-
-		/// <summary>Represent the exclusive end index of the Range.</summary>
-		public Index End { get; }
-
-		/// <summary>Construct a Range object using the start and end indexes.</summary>
-		/// <param name="start">Represent the inclusive start index of the range.</param>
-		/// <param name="end">Represent the exclusive end index of the range.</param>
-		public Range(Index start, Index end)
-		{
-			Start = start;
-			End = end;
-		}
-
-		/// <summary>Indicates whether the current Range object is equal to another object of the same type.</summary>
-		/// <param name="value">An object to compare with this object</param>
-		public override bool Equals([NotNullWhen(true)] object? value) =>
-			value is Range r &&
-			r.Start.Equals(Start) &&
-			r.End.Equals(End);
-
-		/// <summary>Indicates whether the current Range object is equal to another Range object.</summary>
-		/// <param name="other">An object to compare with this object</param>
-		public bool Equals(Range other) => other.Start.Equals(Start) && other.End.Equals(End);
-
-		/// <summary>Returns the hash code for this instance.</summary>
-		public override int GetHashCode()
-		{
-#if (!NETSTANDARD2_0 && !NETFRAMEWORK)
-            return HashCode.Combine(Start.GetHashCode(), End.GetHashCode());
-#else
-			//https://github.com/dotnet/runtime/blob/v9.0.0/src/libraries/System.Private.CoreLib/src/System/Numerics/Hashing/HashHelpers.cs
-			var h1 = Start.GetHashCode();
-			var h2 = End.GetHashCode();
-			// RyuJIT optimizes this to use the ROL instruction
-			// Related GitHub pull request: https://github.com/dotnet/coreclr/pull/1830
-			uint rol5 = ((uint)h1 << 5) | ((uint)h1 >> 27);
-			return ((int)rol5 + h1) ^ h2;
-#endif
-		}
-
-		/// <summary>Converts the value of the current Range object to its equivalent string representation.</summary>
-		public override string ToString()
-		{
-#if (!NETSTANDARD2_0 && !NETFRAMEWORK)
-            Span<char> span = stackalloc char[2 + (2 * 11)]; // 2 for "..", then for each index 1 for '^' and 10 for longest possible uint
-            int pos = 0;
-
-            if (Start.IsFromEnd)
-            {
-                span[0] = '^';
-                pos = 1;
-            }
-            bool formatted = ((uint)Start.Value).TryFormat(span.Slice(pos), out int charsWritten);
-            Debug.Assert(formatted);
-            pos += charsWritten;
-
-            span[pos++] = '.';
-            span[pos++] = '.';
-
-            if (End.IsFromEnd)
-            {
-                span[pos++] = '^';
-            }
-            formatted = ((uint)End.Value).TryFormat(span.Slice(pos), out charsWritten);
-            Debug.Assert(formatted);
-            pos += charsWritten;
-
-            return new string(span.Slice(0, pos));
-#else
-			return Start.ToString() + ".." + End.ToString();
-#endif
-		}
-
-		/// <summary>Create a Range object starting from start index to the end of the collection.</summary>
-		public static Range StartAt(Index start) => new Range(start, Index.End);
-
-		/// <summary>Create a Range object starting from first element in the collection to the end Index.</summary>
-		public static Range EndAt(Index end) => new Range(Index.Start, end);
-
-		/// <summary>Create a Range object starting from first element to the end.</summary>
-		public static Range All => new Range(Index.Start, Index.End);
-
-		/// <summary>Calculate the start offset and length of range object using a collection length.</summary>
-		/// <param name="length">The length of the collection that the range will be used with. length has to be a positive value.</param>
-		/// <remarks>
-		/// For performance reason, we don't validate the input length parameter against negative values.
-		/// It is expected Range will be used with collections which always have non negative length/count.
-		/// We validate the range is inside the length scope though.
-		/// </remarks>
-		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		public (int Offset, int Length) GetOffsetAndLength(int length)
-		{
-			int start = Start.GetOffset(length);
-			int end = End.GetOffset(length);
-
-			if ((uint)end > (uint)length || (uint)start > (uint)end)
-			{
-				ThrowArgumentOutOfRangeException();
-			}
-
-			return (start, end - start);
-		}
-
-		private static void ThrowArgumentOutOfRangeException()
-		{
-			throw new ArgumentOutOfRangeException("length");
-		}
-	}
-}
-
-#endif
